# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AdminerSessionResponse": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "AdminerSessionResponse",
        "type": "object"
      },
...
      "CreateDatabaseRequest": {
        "properties": {
          "name": {
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z][a-z0-9_]{1,63}$",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateDatabaseRequest",
        "type": "object"
      },
...
      "CredentialsResponse": {
        "properties": {
          "database": {
            "$ref": "#/components/schemas/DatabaseResponse"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "database",
          "password"
        ],
        "title": "CredentialsResponse",
        "type": "object"
      },
      "DatabaseListEntry": {
        "properties": {
          "database": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatabaseResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "missing": {
            "default": false,
            "title": "Missing",
            "type": "boolean"
          },
          "orphan_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orphan Name"
          },
          "site_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Domain"
          }
        },
        "title": "DatabaseListEntry",
        "type": "object"
      },
      "DatabaseResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "db_user": {
            "title": "Db User",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "site_id": {
            "title": "Site Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "site_id",
          "name",
          "db_user",
          "purpose",
          "created_at"
        ],
        "title": "DatabaseResponse",
        "type": "object"
      },
...
      "DeleteDatabaseRequest": {
        "properties": {
          "confirm_name": {
            "title": "Confirm Name",
            "type": "string"
          }
        },
        "required": [
          "confirm_name"
        ],
        "title": "DeleteDatabaseRequest",
        "type": "object"
      },
...
      "DeleteSiteRequest": {
        "properties": {
          "confirm_domain": {
            "title": "Confirm Domain",
            "type": "string"
          }
        },
        "required": [
          "confirm_domain"
        ],
        "title": "DeleteSiteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationAccepted": {
        "properties": {
          "operation_id": {
            "title": "Operation Id",
            "type": "integer"
          },
          "site": {
            "$ref": "#/components/schemas/SiteResponse"
          }
        },
        "required": [
          "site",
          "operation_id"
        ],
        "title": "OperationAccepted",
        "type": "object"
      },
...
      "SiteResponse": {
        "properties": {
          "behind_cloudflare": {
            "title": "Behind Cloudflare",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "doc_root": {
            "title": "Doc Root",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "php_memory_limit": {
            "title": "Php Memory Limit",
            "type": "string"
          },
          "php_upload_max_filesize": {
            "title": "Php Upload Max Filesize",
            "type": "string"
          },
          "php_version": {
            "title": "Php Version",
            "type": "string"
          },
          "site_user": {
            "title": "Site User",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "wordpress": {
            "title": "Wordpress",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "domain",
          "site_user",
          "doc_root",
          "php_version",
          "status",
          "error_message",
          "php_memory_limit",
          "php_upload_max_filesize",
          "wordpress",
          "behind_cloudflare",
          "created_at"
        ],
        "title": "SiteResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "HostyPanel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/databases": {
      "get": {
        "operationId": "list_databases_api_databases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatabaseListEntry"
                  },
                  "title": "Response List Databases Api Databases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Databases",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/databases/adminer-session": {
      "post": {
        "description": "Mint a short-lived ticket; the /adminer proxy swaps it for a cookie.\n\nAdmin-only: Adminer's login form takes any server credentials, so the\nproxy must not be reachable by client tenants (Phase 11a).",
        "operationId": "adminer_session_api_databases_adminer_session_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminerSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Adminer Session",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/databases/orphans/{name}": {
      "delete": {
        "description": "Drop a physical database that has no panel record (shown as 'orphan').\n\nGuards: never a system schema, never a panel-managed database (those go\nthrough their own delete, which also removes the DB user), must actually\nexist, and the name must be typed back to confirm.",
        "operationId": "delete_orphan_database_api_databases_orphans__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Orphan Database",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/databases/sites/{site_id}": {
      "post": {
        "operationId": "create_database_api_databases_sites__site_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Database",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/databases/{database_id}": {
      "delete": {
        "operationId": "delete_database_api_databases__database_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "title": "Database Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Database",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/databases/{database_id}/reset-password": {
      "post": {
        "operationId": "reset_password_api_databases__database_id__reset_password_post",
        "parameters": [
          {
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "title": "Database Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Password",
        "tags": [
          "databases"
        ]
      }
    },
...
    "/api/sites/{site_id}": {
      "delete": {
        "operationId": "delete_site_api_sites__site_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Site",
        "tags": [
          "sites"
        ]
      },
      "get": {
        "operationId": "get_site_api_sites__site_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Site",
        "tags": [
          "sites"
        ]
      }
    },
...
  }
}